LPS-44960 refactor OpenIdUtil to remove unnecessary util methods and convert to osgi integration point
