+++ portal-impl/src/com/liferay/portal/util/OpenIdUtil.java
@@ -14,82 +14,37 @@
 
 package com.liferay.portal.util;
 
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.util.Http;
-import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
+import aQute.bnd.annotation.ProviderType;
 
-import org.openid4java.consumer.ConsumerException;
-import org.openid4java.consumer.ConsumerManager;
-import org.openid4java.consumer.InMemoryConsumerAssociationStore;
-import org.openid4java.consumer.InMemoryNonceVerifier;
+import com.liferay.portal.kernel.openid.OpenId;
+import com.liferay.registry.Registry;
+import com.liferay.registry.RegistryUtil;
+import com.liferay.registry.ServiceTracker;
 
 /**
  * @author Jorge Ferrer
  */
+@ProviderType
 public class OpenIdUtil {
 
-	public static ConsumerManager getConsumerManager() {
-		_instance._initialize();
-
-		return _instance._manager;
-	}
-
-	public static String getScreenName(String openId) {
-		String result = normalize(openId);
-
-		if (result.startsWith(Http.HTTP_WITH_SLASH)) {
-			result = result.substring(Http.HTTP_WITH_SLASH.length());
-		}
-
-		if (result.startsWith(Http.HTTPS_WITH_SLASH)) {
-			result = result.substring(Http.HTTPS_WITH_SLASH.length());
-		}
-
-		result = StringUtil.replace(
-			result, new String[] {StringPool.SLASH, StringPool.UNDERLINE},
-			new String[] {StringPool.PERIOD, StringPool.PERIOD});
-
-		return result;
-	}
-
 	public static boolean isEnabled(long companyId) {
-		return PrefsPropsUtil.getBoolean(
-			companyId, PropsKeys.OPEN_ID_AUTH_ENABLED,
-			PropsValues.OPEN_ID_AUTH_ENABLED);
+		return getOpenId().isEnabled(companyId);
 	}
 
-	public static String normalize(String identity) {
-		String result = identity;
-
-		if (result.endsWith(StringPool.SLASH)) {
-			result = result.substring(0, result.length() - 1);
-		}
-
-		return result;
+	protected static OpenId getOpenId() {
+		return _instance._serviceTracker.getService();
 	}
 
-	private void _initialize() {
-		try {
-			if (_manager == null) {
-				_manager = new ConsumerManager();
+	private OpenIdUtil() {
+		Registry registry = RegistryUtil.getRegistry();
 
-				_manager.setAssociations(
-					new InMemoryConsumerAssociationStore());
-				_manager.setNonceVerifier(new InMemoryNonceVerifier(5000));
-			}
-		}
-		catch (ConsumerException ce) {
-			_log.error(ce.getMessage());
-		}
+		_serviceTracker = registry.trackServices(OpenId.class);
+
+		_serviceTracker.open();
 	}
-
-	private static final Log _log = LogFactoryUtil.getLog(OpenIdUtil.class);
 
 	private static final OpenIdUtil _instance = new OpenIdUtil();
 
-	private ConsumerManager _manager;
+	private final ServiceTracker<OpenId, OpenId> _serviceTracker;
 
 }
